requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.9
django-cors-middleware==1.5.0
django-rest-registration==0.5.3
django-sendgrid-v5==0.8.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.4.0
future==0.18.2
gunicorn==20.0.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2==2.8.4
PyJWT==1.7.1
python-decouple==3.3
python-http-client==3.2.1
pytz==2019.3
requests==2.22.0
sendgrid==6.1.0
sqlparse==0.3.0
typing==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==5.0.1
